Phase C-1 cascade: move ResumeVirtualLayoutSummary to majit-ir::resumedata
Serialization-friendly summary of VirtualInfo carrying the same shape
but with ResumeValueLayoutSummary instead of ResumeValueSource field
values.  Depends only on Const/Type/DescrRef/FieldDescrInfo/ArrayDescrInfo
(all in majit-ir) plus already-moved ResumeValueLayoutSummary.

Two inherent impl methods (to_virtual_info, to_exit_virtual_layout)
cannot follow due to cross-crate orphan rule — they reference
VirtualInfo (still in metainterp) and ExitVirtualLayout (in
majit-backend).  Extract them to free functions
(resume_virtual_layout_to_virtual_info /
resume_virtual_layout_to_exit_virtual_layout) in resume.rs; only
two internal callsites need updating to free-fn syntax.

Re-exported from majit-metainterp::resume for caller compatibility.
